requirements.txt
flask
faiss-cpu
numpy
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
langchain-ollama
sentence-transformers
pypdf
